requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.2
django-filter==23.3
idna==3.4
oauthlib==3.2.2
pip==23.3.1
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7

